requirements.txt
openai
sounddevice
numpy
scipy
pynput
python-dotenv
